+++ src/main/java/com/ty/Shopit/Repo/BrandRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ty.Shopit.Entity.Brand;
 
-public interface BrandRepo extends JpaRepository<Brand, Integer>{
+public interface BrandRepo extends JpaRepository<Brand, Long>{
 
 }
