+++ src/main/java/com/ty/Shopit/Repo/CustomerProductRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ty.Shopit.Entity.CustomerProduct;
 
-public interface CustomerProductRepo extends JpaRepository<CustomerProduct, Integer>{
+public interface CustomerProductRepo extends JpaRepository<CustomerProduct, Long>{
 
 }
